refactor: 💡 Fix Module component name
