+++ apps/e-learning-app/src/app/app.js
@@ -1,12 +1,17 @@
 import styles from './app.module.css';
-import { ReactComponent as Logo } from './logo.svg';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
-import { Route, Link } from 'react-router-dom';
-import { environment } from '../environments/environment';
-import Module from './components/Model';
+import Module from './components/Module';
 
 export function App() {
   return (
+  <div className={styles.app}>
+    <Router>
+      <Switch>
+        <Route exact path={'/module'} component={Module} />
+      </Switch>
+    </Router>
+  </div>
   )
 }
 export default App;
