+++ apps/e-learning-app/src/app/app.module.css
@@ -1,4 +1,4 @@
-.app {
+/* .app {
   font-family: sans-serif;
   min-width: 300px;
   max-width: 600px;
@@ -125,4 +125,4 @@
   height: 16px;
   width: 16px;
   margin-right: 4px;
-}
+} */
