+++ apps/e-learning-app/src/app/components/Module.js
